refactor: move onboarding forms folders to OnboardingPage folder
