+++ src/pages/OnboardingPage/index.tsx
@@ -7,12 +7,12 @@ import { Box } from '@mui/system';
 import ArrowLeftIcon from 'src/assets/icons/arrow-left.svg';
 import { onboardingSteps } from 'src/common/constants';
 import Header from 'src/components/Header';
-import OnboardingCreditCardForm from 'src/components/OnboardingCreditCardForm';
-import OnboardingHeaderItem from 'src/components/OnboardingHeaderItem';
-import OnboardingInfoForm from 'src/components/OnboardingInfoForm';
-import OnboardingRoleForm from 'src/components/OnboardingRoleForm';
-import OnboardingShippingForm from 'src/components/OnboardingShippingForm';
-import OnboardingSizeForm from 'src/components/OnboardingSizeForm';
+import OnboardingCreditCardForm from 'src/pages/OnboardingPage/CreditCardForm';
+import OnboardingHeaderItem from 'src/pages/OnboardingPage/HeaderItem';
+import OnboardingInfoForm from 'src/pages/OnboardingPage/InfoForm';
+import OnboardingRoleForm from 'src/pages/OnboardingPage/RoleForm';
+import OnboardingShippingForm from 'src/pages/OnboardingPage/ShippingForm';
+import OnboardingSizeForm from 'src/pages/OnboardingPage/SizeForm';
 import { useAppSelector } from 'src/redux/hooks';
 import { selectOnboardingStep } from 'src/redux/user/userSlice';
 import theme from 'src/theme';
